requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
discord.py==0.16.12
idna==2.7
multidict==4.5.2
requests==2.20.0
urllib3==1.24.1
websockets==3.4
psycopg2==2.6.2
